ai-loan-processing/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.1
python-dotenv==1.0.1
Jinja2==3.1.4
Pillow==10.4.0
pytesseract==0.3.10
pdf2image==1.17.0
opencv-python-headless==4.10.0.84
openai==1.35.10
streamlit==1.36.0
sqlalchemy==2.0.31
pytest==8.3.1
